# Improve performance of compareTo