This PR improves performance of `compareTo`. This is accomplished by saving the expensive call to `compareUnsignedLongs` for longs that are know to be equal.
This saves 0 to 40 percent of time, depending on the joint portion of the slices.