@martint 2016-08-09
Can you add a JMH benchmark?