@fmeiser 2016-08-09
@martint Provided the class for benchmark and the results. I have also added a commit with direct use of address with unsafe.

Original Version
```
Benchmark                 Mode  Cnt   Score   Error  Units
BenchmarkSlice.compareTo  avgt   20  32,132 ± 0,044  ns/op
```

Reorded compare
```
Benchmark                 Mode  Cnt   Score   Error  Units
BenchmarkSlice.compareTo  avgt   20  28,729 ± 0,323  ns/op
```

Using directly address with unsafe
```
Benchmark                 Mode  Cnt   Score   Error  Units
BenchmarkSlice.compareTo  avgt   20  26,221 ± 0,240  ns/op
```