@martint 2016-08-12
I made a few changes and merged it. Thanks!

BTW, here's a more comprehensive benchmark result for different sizes:

```
Before:

Benchmark                 (size)  Mode  Cnt     Score     Error  Units
BenchmarkSlice.compareTo       1  avgt   20     6.248 ±   0.535  ns/op
BenchmarkSlice.compareTo       7  avgt   20    14.806 ±   0.233  ns/op
BenchmarkSlice.compareTo       8  avgt   20     5.967 ±   0.098  ns/op
BenchmarkSlice.compareTo      16  avgt   20    10.031 ±   0.171  ns/op
BenchmarkSlice.compareTo      32  avgt   20    15.668 ±   0.295  ns/op
BenchmarkSlice.compareTo      64  avgt   20    24.333 ±   0.675  ns/op
BenchmarkSlice.compareTo     127  avgt   20    50.837 ±   0.741  ns/op
BenchmarkSlice.compareTo   32779  avgt   20  8359.595 ± 122.491  ns/op

After:

Benchmark                 (size)  Mode  Cnt     Score    Error  Units
BenchmarkSlice.compareTo       1  avgt   20     6.036 ±  0.097  ns/op
BenchmarkSlice.compareTo       7  avgt   20    11.913 ±  0.195  ns/op
BenchmarkSlice.compareTo       8  avgt   20     6.176 ±  0.133  ns/op
BenchmarkSlice.compareTo      16  avgt   20     7.561 ±  0.128  ns/op
BenchmarkSlice.compareTo      32  avgt   20     9.654 ±  0.132  ns/op
BenchmarkSlice.compareTo      64  avgt   20    14.359 ±  0.361  ns/op
BenchmarkSlice.compareTo     127  avgt   20    27.731 ±  0.361  ns/op
BenchmarkSlice.compareTo   32779  avgt   20  3666.218 ± 70.419  ns/op
```